python_backend/requirements.txt

flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.0.1
supabase==1.1.1
gunicorn==21.2.0
requests==2.31.0
boto3==1.28.62
pydantic==2.4.2
python-jose==3.3.0
nltk==3.8.1
scikit-learn==1.3.1
pandas==2.1.1
firecrawl==1.0.0
